+++ Courier/src/utopia/courier/controller/read/EmailReader.scala
@@ -1,23 +1,27 @@
 package utopia.courier.controller.read
 
-import utopia.courier.model.{Email, EmailAddress}
 import utopia.courier.model.read.DeletionRule.NeverDelete
-import utopia.courier.model.read.{DeletionRule, FolderPath, ReadSettings}
+import utopia.courier.model.read.{DeletableEmail, DeletionRule, FolderPath, ReadSettings}
+import utopia.courier.model.{Email, EmailAddress}
 import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.collection.immutable.caching.LazyTree
+import utopia.flow.collection.mutable.VolatileList
 import utopia.flow.operator.EqualsExtensions._
 import utopia.flow.parse.AutoClose._
 import utopia.flow.parse.AutoCloseWrapper
 import utopia.flow.util.StringExtensions._
+import utopia.flow.util.logging.{CollectSingleFailureLogger, Logger}
 import utopia.flow.view.immutable.caching.Lazy
 import utopia.flow.view.mutable.caching.ResettableLazy
+import utopia.flow.view.mutable.eventful.Flag
+import utopia.flow.view.template.eventful.{ChangingWrapper, FlagLike}
 
 import java.io.InputStream
 import java.nio.file.Path
 import java.util.Properties
 import javax.mail.Message.RecipientType
-import javax.mail.internet.MimeMessage
 import javax.mail._
+import javax.mail.internet.MimeMessage
 import scala.annotation.tailrec
 import scala.concurrent.{ExecutionContext, Future}
 import scala.io.Codec
@@ -39,6 +43,15 @@ object EmailReader
 	  */
 	def default(implicit settings: ReadSettings) =
 		apply { headers => new EmailBuilder(headers.toHeaders) }
+	/**
+	 * Creates a new email reader implementation that uses the default email builder,
+	 * including support for pointer-based message deletion.
+	 * NB: Attachments won't be processed
+	 * @param settings Implicit email read settings to use
+	 * @return A new reader
+	 */
+	def deletable(implicit settings: ReadSettings) =
+		withDeletionFlags { (hv, flag) => new EmailBuilder(hv.toHeaders).mapResult { new DeletableEmail(_, flag) } }
 	
 	
 	// OTHER    ------------------------------
@@ -52,7 +65,6 @@ object EmailReader
 	  */
 	def apply[A](makeBuilder: LazyEmailHeadersView => FromEmailBuilder[A])(implicit settings: ReadSettings) =
 		filtered { headers => Some(makeBuilder(headers)) }
-	
 	/**
 	  * Creates a new mail reader that skips some of the incoming messages
 	  * @param makeBuilder A function for creating a new mail processor. Accepts email subject and headers.
@@ -62,7 +74,31 @@ object EmailReader
 	  * @return A new email reader
 	  */
 	def filtered[A](makeBuilder: LazyEmailHeadersView => Option[FromEmailBuilder[A]])(implicit settings: ReadSettings) =
-		new EmailReader[A](settings, makeBuilder)
+		new EmailReader[A](settings, (hv, _) => makeBuilder(hv))
+	/**
+	 * Creates a new email reader that supports message deletion -flags,
+	 * as well as header-based filtering.
+	 * @param makeBuilder A function that generates the used builder from email message headers.
+	 *                    Also accepts generated message deletion flags.
+	 *                    Yields None for emails that should not be processed.
+	 * @param settings Implicit email read settings to use
+	 * @tparam A Type of processing result
+	 * @return A new email reader
+	 */
+	def filteredWithDeletionFlags[A](makeBuilder: (LazyEmailHeadersView, Flag) => Option[FromEmailBuilder[A]])
+	                                (implicit settings: ReadSettings) =
+		new EmailReader[A](settings, (hv, df) => makeBuilder(hv, df.get), generateDeletionFlags = true)
+	/**
+	 * Creates a new email reader that supports message deletion -flags.
+	 * @param makeBuilder A function that generates the used builder from email message headers.
+	 *                     Also accepts generated message deletion flags.
+	 * @param settings Implicit email read settings to use
+	 * @tparam A Type of processing result
+	 * @return A new email reader
+	 */
+	def withDeletionFlags[A](makeBuilder: (LazyEmailHeadersView, Flag) => FromEmailBuilder[A])
+	                        (implicit settings: ReadSettings) =
+		filteredWithDeletionFlags { (hv, df) => Some(makeBuilder(hv, df)) }
 	
 	/**
 	  * @param attachmentsStoreDirectory Directory where the read attachments will be stored
@@ -71,7 +107,6 @@ object EmailReader
 	  */
 	def defaultWithAttachments(attachmentsStoreDirectory: Path)(implicit settings: ReadSettings) =
 		apply { headers => new EmailBuilder(headers.toHeaders, Some(attachmentsStoreDirectory)) }
-	
 	/**
 	  * Creates a new mail reader that filters based on email subject
 	  * @param filter A function that accepts email headers and returns whether they should be processed
@@ -85,7 +120,6 @@ object EmailReader
 			else
 				None
 		}
-	
 	/**
 	  * Creates a new mail reader that filters based on email subject and stores attachments
 	  * @param filter A function that accepts email headers and returns whether they should be processed
@@ -101,6 +135,38 @@ object EmailReader
 			else
 				None
 		}
+	
+	/**
+	 * Creates a new email reader that uses filtering and pointer-based message deletion
+	 * @param filter A filtering function to use. Accepts email headers.
+	 * @param settings Implicit email read settings to use
+	 * @return A new email reader
+	 */
+	def filteredAndDeletable(filter: LazyEmailHeadersView => Boolean)
+	                        (implicit settings: ReadSettings): EmailReader[DeletableEmail[Email]] =
+		filteredWithDeletionFlags { (headers, flag) =>
+			if (filter(headers))
+				Some(new EmailBuilder(headers.toHeaders).mapResult { new DeletableEmail(_, flag) })
+			else
+				None
+		}
+	/**
+	 * Creates a new email reader that uses filtering, pointer-based message deletion and attachment-storing
+	 * @param attachmentsDirectory Directory where attachment files should be stored
+	 * @param filter   A filtering function to use. Accepts email headers.
+	 * @param settings Implicit email read settings to use
+	 * @return A new email reader
+	 */
+	def filteredAndDeletableDefaultWithAttachments(attachmentsDirectory: Path)
+	                                              (filter: LazyEmailHeadersView => Boolean)
+	                                              (implicit settings: ReadSettings): EmailReader[DeletableEmail[Email]] =
+		filteredWithDeletionFlags { (headers, flag) =>
+			if (filter(headers))
+				Some(new EmailBuilder(headers.toHeaders, Some(attachmentsDirectory))
+					.mapResult { new DeletableEmail(_, flag) })
+			else
+				None
+		}
 }
 
 /**
@@ -108,7 +174,9 @@ object EmailReader
   * @author Mikko Hilpinen
   * @since 11.9.2021, v0.1
   */
-class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView => Option[FromEmailBuilder[A]])
+class EmailReader[A](settings: ReadSettings,
+                     makeBuilder: (LazyEmailHeadersView, Option[Flag]) => Option[FromEmailBuilder[A]],
+                     generateDeletionFlags: Boolean = false)
 {
 	/**
 	  * Asynchronously reads and parses email
@@ -132,26 +200,38 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 	  * @param f Function that accepts an iterator that returns read messages.
 	 *                   The specified iterator may contain any number of failures.
 	  * @param exc Implicit execution context
-	  * @return A future with the read results, when they come available
+	  * @param log A logging implementation used to record errors that occur outside the iteration process
+	 * @return A future with the read results, when they come available
 	  */
 	def iterateAsync[B](targetFolders: TargetFolders = TargetFolders.inbox, skipMessageCount: Int = 0,
-	                    deletionRule: DeletionRule = NeverDelete)(f: Iterator[Try[A]] => B)
-	                   (implicit exc: ExecutionContext) =
+	                    deletionRule: DeletionRule = NeverDelete)
+	                   (f: Iterator[Try[A]] => B)
+	                   (implicit exc: ExecutionContext, log: Logger) =
 		Future { iterateBlocking(targetFolders, skipMessageCount, deletionRule)(f) }
 	
 	/**
 	  * Reads email data from the targeted message folder.
+	 *  Interrupts the reading process if any failure is encountered
 	  * @param targetFolders Targeted folder or folders. Default = Inbox.
 	  * @param maxMessageCount Maximum number of read messages (includes those skipped).
 	  *                        Negative if not limited (default)
 	  * @param skipMessageCount Number of messages to skip from the beginning (default = 0)
 	  * @param deletionRule A rule to apply to deleting messages (default = never delete messages)
-	  * @return Newly read data. Failure if connection or read setup failed
+	 * @return Newly read data. Failure if a failure was encountered.
 	  */
+	// TODO: Add a version that returns TryCatch and doesn't interrupt on a read failure
 	def readBlocking(targetFolders: TargetFolders = TargetFolders.inbox,
 	                 maxMessageCount: Int = -1, skipMessageCount: Int = 0,
 	                 deletionRule: DeletionRule = NeverDelete) =
-		iterateBlocking(targetFolders, skipMessageCount, deletionRule) { _.take(maxMessageCount).toTry }
+	{
+		// Catches certain failures using a special logger
+		implicit val log: CollectSingleFailureLogger = CollectSingleFailureLogger.ignoringMessages()
+		// Performs the iteration
+		val primaryResult = iterateBlocking(targetFolders, skipMessageCount,
+			deletionRule) { _.take(maxMessageCount).toTry }
+		// Fails if an error was logged
+		primaryResult.failIf(log.value)
+	}
 	
 	/**
 	  * Reads and processes email data using a message iterator. Message reading is terminated if
@@ -161,11 +241,14 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 	  * @param deletionRule A rule to apply to deleting messages (default = never delete messages)
 	  * @param f Function that accepts an iterator that returns read messages.
 	 *          The specified iterator may contain any number of failures.
-	  * @return Newly read data. Failure if connection or read setup failed
+	  * @param log Logging implementation used for recording errors encountered after the iteration has completed,
+	 *            or upon iteration finalization
+	 * @return Newly read data. Failure if connection or read setup failed
 	  */
 	def iterateBlocking[B](targetFolders: TargetFolders = TargetFolders.inbox,
-	                       skipMessageCount: Int = 0,
-	                       deletionRule: DeletionRule = NeverDelete)(f: Iterator[Try[A]] => B): B =
+	                       skipMessageCount: Int = 0, deletionRule: DeletionRule = NeverDelete)
+	                      (f: Iterator[Try[A]] => B)
+	                      (implicit log: Logger): B =
 	{
 		// Initializes the session
 		settings.modify(new Properties).flatMap { properties =>
@@ -224,7 +307,8 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 	// NESTED   -------------------------------------
 	
 	private class FoldersIterator(source: Iterator[Try[Folder]], skipMessageCount: Int, deletionRule: DeletionRule)
-		extends Iterator[Try[Message]]
+	                             (implicit log: Logger)
+		extends Iterator[Try[(Message, Option[Flag])]]
 	{
 		// ATTRIBUTES   ---------------------------
 		
@@ -236,6 +320,11 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		private var nextMessageIndex = 1
 		private var queuedFailures: Iterator[Throwable] = Iterator.empty
 		
+		private val closedFlag = Flag()
+		
+		// Sometimes message deletions are queued to be completed later
+		private lazy val deletionQueuePointer = VolatileList[(Folder, Message)]()
+		
 		override def hasNext: Boolean = queuedFailures.hasNext || openFolder.exists { _._2 > nextMessageIndex } ||
 			openNextFolder()
 		
@@ -243,14 +332,24 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		// IMPLEMENTED  --------------------------
 		
 		@tailrec
-		override final def next(): Try[Message] = queuedFailures.nextOption() match {
+		override final def next(): Try[(Message, Option[Flag])] = queuedFailures.nextOption() match {
 			// Case: A failure was queued => Returns the queued failure
 			case Some(error) => Failure(error)
 			case None =>
 				openFolder.filter { _._2 > nextMessageIndex } match {
 					// Case: A folder has been opened (as expected) => Reads the next message from the folder
 					case Some((folder, _)) =>
-						val message = Try { folder.getMessage(nextMessageIndex) }
+						val message = Try {
+							val message = folder.getMessage(nextMessageIndex)
+							// May generate a deletion flag for the message as well
+							val deletionFlag = {
+								if (generateDeletionFlags)
+									Some(new DeleteMessagePointer(folder, message, queueMessageDeletion))
+								else
+									None
+							}
+							(message, deletionFlag)
+						}
 						nextMessageIndex += 1
 						// Returns the message read result
 						message
@@ -272,72 +371,126 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		 * @return Closes the currently open folder, if applicable.
 		 *         Should be called before discarding this iterator, unless it has already been consumed
 		 */
-		def close(): Try[Unit] = openFolder match {
-			case Some((folder, _)) => close(folder)
-			case None => Success(())
+		def close(): Unit = {
+			// Performs queued message deletions, if there were any
+			unqueueDeletions()
+			// Closes the currently open folder
+			// Catches and logs close failures
+			openFolder.foreach { case (folder, _) =>
+				close(folder).logFailureWithMessage("Failed to close the last folder")
+			}
+			// Logs errors that were queued but not processed
+			queuedFailures.foreach { log(_) }
+		}
+		
+		private def queueMessageDeletion(folder: Folder, message: Message): Unit = {
+			// Case: Iteration has already completed =>
+			//       Immediately deletes the messages (requires folder-opening & closing)
+			if (closedFlag.value)
+				AutoCloseWrapper(folder) { close(_) }.tryConsumeContent { folder =>
+					folder.open(Folder.READ_WRITE)
+					delete(message)
+				}.logFailureWithMessage("Message deletion failed")
+			// Case: Iteration is yet to complete =>
+			//       Queues the deletion to occur later (possibly avoiding unnecessary folder-openings)
+			else
+				deletionQueuePointer :+= (folder -> message)
 		}
 		
 		// Returns if this iterator may return new items afterwards
 		@tailrec
 		private def openNextFolder(): Boolean = {
-			// Closes the previously open folder before opening the next folder
-			openFolder.flatMap { case (folder, _) => close(folder).failure } match {
-				// Case: Failed to close the folder => Queues the failure
-				case Some(failure) =>
-					queue(failure)
-					true
-				// Case: Previous folder closed => Opens the next folder
-				case None =>
-					source.nextOption() match {
-						// Case: Failed to read the next folder => Queues the encountered failure
-						case Some(Failure(error)) =>
-							queue(error)
-							true
-						// Case: Next folder is available
-						case Some(Success(folder)) =>
-							// Attempts to open the folder
-							Try { folder.open(if (deletionRule.canDelete) Folder.READ_WRITE else Folder.READ_ONLY) }
+			// Processes queued deletions, if there are any
+			// If the process throws, returns the failures as next items
+			if (unqueueDeletions())
+				true
+			else {
+				// Closes the previously open folder before opening the next folder
+				openFolder.flatMap { case (folder, _) => close(folder).failure } match {
+					// Case: Failed to close the folder => Queues the failure
+					case Some(failure) =>
+						queue(failure)
+						true
+					// Case: Previous folder closed => Opens the next folder
+					case None =>
+						source.nextOption() match {
+							// Case: Failed to read the next folder => Queues the encountered failure
+							case Some(Failure(error)) =>
+								queue(error)
+								true
+							// Case: Next folder is available
+							case Some(Success(folder)) =>
+								val openMode = {
+									if (deletionRule.canDelete || generateDeletionFlags)
+										Folder.READ_WRITE
+									else
+										Folder.READ_ONLY
+								}
+								// Attempts to open the folder
 								// Counts the messages in the folder
-								.flatMap { _ => Try { folder.getMessageCount } } match {
-								// Case: Folder successfully opened
-								case Success(messageCount) =>
-									// Case: All messages in this folder are requested to be skipped
-									// => Closes the folder immediately
-									if (remainingSkipCount >= messageCount) {
-										remainingSkipCount -= messageCount
-										close(folder) match {
-											// Case: Closing succeeded => Moves to the next folder, if possible
-											case Success(_) => openNextFolder()
-											// Case: Closing failed => Prepares the closing error as the next element
-											case Failure(error) =>
-												queue(error)
-												true
+								Try { folder.open(openMode) }.flatMap { _ => Try { folder.getMessageCount } } match {
+									// Case: Folder successfully opened
+									case Success(messageCount) =>
+										// Case: All messages in this folder are requested to be skipped
+										// => Closes the folder immediately
+										if (remainingSkipCount >= messageCount) {
+											remainingSkipCount -= messageCount
+											close(folder) match {
+												// Case: Closing succeeded => Moves to the next folder, if possible
+												case Success(_) => openNextFolder()
+												// Case: Closing failed => Prepares the closing error as the next element
+												case Failure(error) =>
+													queue(error)
+													true
+											}
 										}
-									}
-									// Case: There are messages to iterate => Prepares for the iteration
-									else {
-										nextMessageIndex = 1 + remainingSkipCount
-										remainingSkipCount = 0
-										openFolder = Some(folder -> messageCount)
+										// Case: There are messages to iterate => Prepares for the iteration
+										else {
+											nextMessageIndex = 1 + remainingSkipCount
+											remainingSkipCount = 0
+											openFolder = Some(folder -> messageCount)
+											true
+										}
+									// Case: Failed to open the folder => Prepares the error as the next element
+									case Failure(error) =>
+										queue(error)
 										true
-									}
-								// Case: Failed to open the folder => Prepares the error as the next element
-								case Failure(error) =>
-									queue(error)
-									true
-							}
-						// Case: No more folders available
-						case None => false
-					}
+								}
+							// Case: No more folders available
+							case None => false
+						}
+				}
 			}
 		}
 		
-		private def close(folder: Folder) = {
+		private def close(folder: Folder): Try[Unit] = {
 			openFolder = None
-			Try { folder.close(deletionRule.canDelete) }
+			Try { folder.close(deletionRule.canDelete || generateDeletionFlags) }
 		}
 		
 		private def queue(error: Throwable) = queuedFailures :+= error
+		
+		// Returns whether any failures were encountered
+		private def unqueueDeletions() = {
+			deletionQueuePointer.mutate { queue =>
+				// Deletes contents from one folder at a time
+				val deleteResults = queue.asMultiMap.map { case (folder, messages) =>
+					// Case: Folder is currently open => Simply flags the messages as deleted
+					if (folder.isOpen)
+						Try { messages.foreach(delete) }
+					// Case: Folder is not open => Opens it for the duration of deletion
+					else
+						AutoCloseWrapper(folder) { close(_) }.tryConsumeContent { folder =>
+							folder.open(Folder.READ_WRITE)
+							messages.foreach(delete)
+						}
+				}
+				// Queues the encountered failures to this iterator
+				// It is possible, however, that these won't ever be read
+				deleteResults.iterator.foreach { _.failure.foreach(this.queue) }
+				deleteResults.exists { _.isFailure } -> Vector()
+			}
+		}
 	}
 	
 	private class MessageIterator(source: RawMessageIterator, deletionRule: DeletionRule) extends Iterator[Try[A]]
@@ -432,8 +585,8 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 	}
 	
 	// Returns Success(None) for messages that are skipped
-	private class RawMessageIterator(source: Iterator[Try[Message]], deletionRule: DeletionRule,
-	                                 makeBuilder: LazyEmailHeadersView => Option[FromEmailBuilder[A]])
+	private class RawMessageIterator(source: Iterator[Try[(Message, Option[Flag])]], deletionRule: DeletionRule,
+	                                 makeBuilder: (LazyEmailHeadersView, Option[Flag]) => Option[FromEmailBuilder[A]])
 		extends Iterator[Try[Option[(Message, FromEmailBuilder[A])]]]
 	{
 		// IMPLEMENTED  -----------------------------
@@ -442,7 +595,7 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		
 		override def next() = {
 			// Reads the next message (may fail)
-			source.next().flatMap { message =>
+			source.next().flatMap { case (message, deletionFlag) =>
 				Try {
 					// Reads the header information (lazily)
 					def subject = message.getSubject
@@ -486,7 +639,7 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 					
 					// Creates a builder, if necessary
 					makeBuilder(new LazyEmailHeadersView(sender, subject, messageId, sentTime, recipients, inReplyTo,
-						references, replyTo))
+						references, replyTo), deletionFlag)
 				} match {
 					case Success(builder) =>
 						// May delete skipped messages (ignores failures)
@@ -507,5 +660,39 @@ class EmailReader[A](settings: ReadSettings, makeBuilder: LazyEmailHeadersView =
 		
 		private def normalizeHeaderValue(value: String) =
 			value.stripControlCharacters.trim.notStartingWith("<").notEndingWith(">")
+	}
+	
+	private class DeleteMessagePointer(folder: Folder, message: Message,
+	                                   queueDeletion: (Folder, Message) => Unit)
+	                                  (implicit log: Logger)
+		extends Flag with ChangingWrapper[Boolean]
+	{
+		// ATTRIBUTES   -----------------------
+		
+		private val lazyFlag = Lazy { Flag() }
+		
+		
+		// IMPLEMENTED  -----------------------
+		
+		override protected def wrapped = lazyFlag.value
+		override def view: FlagLike = wrapped.view
+		
+		override def set(): Boolean = {
+			// Case: Message has not yet been deleted (via this pointer)
+			if (lazyFlag.current.forall { _.isNotSet }) {
+				if (folder.isOpen) {
+					if (folder.getMode == Folder.READ_WRITE)
+						delete(message)
+							.logFailureWithMessage("Failed to delete a message upon deletionFlag.set()")
+					else
+						queueDeletion(folder, message)
+				}
+				else
+					queueDeletion(folder, message)
+				lazyFlag.value.set()
+			}
+			else
+				false
+		}
 	}
 }